requirements.txt
streamlit>=1.35
google-genai>=1.0
pypdf>=4.2
